api/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
httpx>=0.27,<1.0
pydantic-settings>=2.6,<3.0
python-jose[cryptography]>=3.3,<4.0
pytest>=8.3,<9.0
pytest-asyncio>=0.24,<1.0
email-validator>=2.2,<3.0
psycopg[binary]>=3.2,<4.0
